delete redundant comment, correct file name: msfnn.py->msffn.py
